requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
snowflake-connector-python>=3.0.0
mcp>=0.9.0
streamlit>=1.28.0
plotly>=5.17.0




